+++ otserv/trunk/vocation.h
@@ -35,8 +35,8 @@ public:
 	const std::string& getName() const {return name;}
 	const std::string& getDescription() const {return description;}
 
-	uint32_t getReqSkillTries(int32_t skill, int32_t level);
-	uint32_t getReqMana(int32_t magLevel);
+	uint64_t getReqSkillTries(int32_t skill, int32_t level);
+	uint64_t getReqMana(int32_t magLevel);
 	uint32_t getHPGain() const {return gainHP;};
 	uint32_t getManaGain() const {return gainMana;};
 	uint32_t getCapGain() const {return gainCap;};
